pyproject.toml
[project]
name = "beeai-agents"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "IBM Corp." }]
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyyaml>=6.0.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "ddgs (>=9.4.3,<10.0.0)",
    "beeai-framework>=0.1.34",
    "beeai-sdk==0.3.0rc7",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[project.scripts]
server = "beeai_agents.agent:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
